Generate a new JAR file - Rename files to fix error in packing javaFX dependencies
